Extract method object for notifying fo form updates
