+++ collect_app/src/main/java/org/odk/collect/android/backgroundwork/SchedulerBackgroundWorkManager.java
@@ -1,7 +1,7 @@
 package org.odk.collect.android.backgroundwork;
 
-import org.odk.collect.android.formmanagement.AutoUpdateTaskSpec;
-import org.odk.collect.android.formmanagement.SyncFormsTaskSpec;
+import org.odk.collect.android.formmanagement.previouslydownloaded.AutoUpdateTaskSpec;
+import org.odk.collect.android.formmanagement.matchexactly.SyncFormsTaskSpec;
 import org.odk.collect.android.upload.AutoSendWorker;
 import org.odk.collect.async.Scheduler;
