+++ collect_app/src/main/java/org/odk/collect/android/notifications/NotificationRepository.java
@@ -1,0 +1,8 @@
+package org.odk.collect.android.notifications;
+
+public interface NotificationRepository {
+
+    void markFormUpdateNotified(String formId, String formHash, String manifestHash);
+
+    boolean hasFormUpdateBeenNotified(String formHash, String manifestHash);
+}
